2024/08/28 管理员登录、获取菜单数据

+++ model/Model.go
@@ -5,8 +5,9 @@ import (
 	"time"
 )
 
+// Model 基模型
 type Model struct {
-	ID        uint           `gorm:"primarykey" json:"id"`
+	ID        uint64         `gorm:"primarykey" json:"id"`
 	CreatedAt time.Time      `json:"created_at"`
 	UpdatedAt time.Time      `json:"updated_at"`
 	DeletedAt gorm.DeletedAt `gorm:"index" json:"deleted_at"`
